Added updated tables and authentication

+++ Controllers/EventController.cs
@@ -22,8 +22,10 @@ public class EventController : Controller
     }
 
     [HttpDelete("DeleteEvent/{Id}")]
-    public async Task<IActionResult> DeleteEvent(int Id)
+    public async Task<IActionResult> DeleteEvent(Guid Id, [FromQuery] Guid UserId)
     {
+        // Check if the user is an admin
+        if(_context.Admins.SingleOrDefault(_=>_.Id == UserId) == null) return Unauthorized();
         if(_context.Events.SingleOrDefault(u => u.Id == Id) == null) return NotFound();
         await _eventservice.DeleteEvent(Id);
         return Ok();
@@ -38,7 +40,7 @@ public class EventController : Controller
     }
 
     [HttpPut("EditEvent/{Id}")]
-    public async Task<IActionResult> EditEvent([FromBody] Events NewEvent, int Id)
+    public async Task<IActionResult> EditEvent([FromBody] Events NewEvent, Guid Id)
     {
         if(_context.Events.SingleOrDefault(u => u.Id == Id) == null) return NotFound();
         if(NewEvent == null) return BadRequest();
